feat: Add a seq2seq with attention trained in Kaggle and modified Sentences Clustering
I have modified some minor features to improve the the results in the Sentence Clustering notebook. And I have include a new
notebook trained in Kagglr with a text summarizer in Pytorch using a seq2seq with attention and force teaching model. But it still
need some modification and improvements, under construction.