feat: A seq2seq with attention and force teaching using fast ai
I have included a notebook with a step by step development of a seq2seq with attention and force teaching using Fast ai. Then,
for training in a custom container on Google Cloud AI Platform, I have created 3 python scrips: data_process, model and task and a
Dockerfile. So, we can create a docker image, publish on GCR (Cloud registry) and run vm with that container for training the
model forsome hours.